delete v2 model, update code and reorganize files
