+++ python/URL Classifier/main.py
@@ -1,4 +1,4 @@
-from verifiers import URLVerifier
+import helper
 
 def main():
     url = "https://uk-covid-19.webredirect.org/to"
@@ -6,10 +6,7 @@ def main():
 
 
 def check_url(url: str):
-    verifier = URLVerifier(url)
-
-    probs = verifier.get_ml_prediction()
-
+    probs = helper.get_model_predictions(url)
     label = "Phishing" if (float(probs[0]) > 0.5) else "Safe"
 
     probs = round((float(probs[0]) * 100), 2)
